Put all test compiler definitions in separate files.
